requirements.txt
absl-py==0.7.1
aniso8601==7.0.0
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-RESTful==0.3.7
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.0.2
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.1.0
numpy==1.16.4
Pillow==6.0.0
protobuf==3.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
scipy==1.3.0
six==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
